Common: moved GraphicSpace to its own header
this is to avoid including GLM where not necessary.
