+++ Solutions/Common.Lib/Common.Lib.vcxproj
@@ -877,6 +877,7 @@
     <ClInclude Include="..\..\Common\gfx\blender.h" />
     <ClInclude Include="..\..\common\gfx\gfx_def.h" />
     <ClInclude Include="..\..\Common\gfx\bitmapdata.h" />
+    <ClInclude Include="..\..\Common\gfx\graphicspace.h" />
     <ClInclude Include="..\..\Common\gfx\image_file.h" />
     <ClInclude Include="..\..\Common\gui\guibutton.h" />
     <ClInclude Include="..\..\Common\gui\guidefines.h" />
