+++ Solutions/Common.Lib/Common.Lib.vcxproj.filters
@@ -1006,5 +1006,8 @@
     <ClInclude Include="..\..\Common\game\roomstruct.h">
       <Filter>Header Files\game</Filter>
     </ClInclude>
+    <ClInclude Include="..\..\Common\gfx\graphicspace.h">
+      <Filter>Header Files\gfx</Filter>
+    </ClInclude>
   </ItemGroup>
 </Project>
